requirements.txt
requests
python-dotenv
python-twitter